requirements.txt
chromadb
sentence_transformers
pandas
openai
python-dotenv
torch
accelerate
deepeval
aiobotocore